+++ superset-frontend/src/explore/components/controls/DateFilterControl/CalendarFrame.tsx
@@ -1,0 +1,57 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+import React from 'react';
+import { t } from '@superset-ui/core';
+import { Radio } from 'src/common/components';
+import { CALENDAR_RANGE_OPTIONS, CALENDAR_RANGE_SET } from './constants';
+import { CalendarRangeType, PreviousCalendarWeek } from './types';
+
+type CalendarFrameProps = {
+  onChange: (timeRange: string) => void;
+  value: string;
+};
+
+export default function CalendarFrame(props: CalendarFrameProps) {
+  function getCalendarRange(value: any): CalendarRangeType {
+    return CALENDAR_RANGE_SET.has(value) ? value : PreviousCalendarWeek;
+  }
+
+  const calendarRange =
+    CALENDAR_RANGE_OPTIONS.find(
+      ({ value }) => value === getCalendarRange(props.value),
+    )?.value || PreviousCalendarWeek;
+
+  return (
+    <>
+      <div className="section-title">
+        {t('Configure Time Range: Previous...')}
+      </div>
+      <Radio.Group
+        value={calendarRange}
+        onChange={(e: any) => props.onChange(e.target.value)}
+      >
+        {CALENDAR_RANGE_OPTIONS.map(({ value, label }) => (
+          <Radio key={value} value={value} className="vertical-radio">
+            {label}
+          </Radio>
+        ))}
+      </Radio.Group>
+    </>
+  );
+}
